Move obfuscated to sdk, make it package protected
